backend/requirements.txt
Flask==2.3.3
pymongo==4.5.0
nltk==3.8.1
flask-cors==4.0.0
python-docx==1.1.2
python-pptx==0.6.23
python-dotenv==1.0.0
PyPDF2==3.0.1
uvicorn